# food_db/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Macro_vision",
        "title": "Macro_vision backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/ingredients": {
            "get": {
                "description": "Retrieve the full list of ingredients from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Get all ingredients",
                "responses": {
                    "200": {
                        "description": "List of ingredients",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Ingredient"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ingredients/search": {
            "get": {
                "description": "Search ingredients by a query string",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Search ingredients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching ingredients",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Ingredient"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ingredients/{ingredient_id}": {
            "get": {
                "description": "Retrieve a single ingredient by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingredients"
                ],
                "summary": "Get ingredient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingredient ID",
                        "name": "ingredient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ingredient",
                        "schema": {
                            "$ref": "#/definitions/database.Ingredient"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.Ingredient": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string"
                },
                "ingredient_id": {
                    "type": "string"
                },
                "ingredient_name": {
                    "type": "string"
                }
            }
        },
...
    }
}